requirements.txt
Flask==3.0.0
opencv-python==4.8.1.78
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
scikit-image==0.22.0
tensorflow==2.16.1
joblib==1.3.2
Pillow==10.1.0
Werkzeug==3.0.0
tqdm==4.66.1
